refactor: centralize TrackMetadata in shared layer
