+++ src/omym/features/metadata/usecases/music_file_processor.py
@@ -31,7 +31,7 @@ from omym.platform.db.db_manager import DatabaseManager
 from omym.platform.logging import logger
 from omym.platform.musicbrainz.client import configure_romanization_cache
 
-from ..domain.track_metadata import TrackMetadata
+from omym.shared.track_metadata import TrackMetadata
 from .directory_runner import run_directory_processing
 from .file_runner import run_file_processing
 from .file_operations import calculate_file_hash, generate_target_path, move_file
