+++ tests/features/path/test_music_file_renamer.py
@@ -7,7 +7,7 @@ from omym.features.path.usecases.renamer import (
     CachedArtistIdGenerator,
     DirectoryGenerator,
 )
-from omym.features.metadata.domain.track_metadata import TrackMetadata
+from omym.shared.track_metadata import TrackMetadata
 from omym.platform.db.cache.artist_cache_dao import ArtistCacheDAO
 from omym.platform.db.db_manager import DatabaseManager
